🚑 Make cruiseControl.targetSpeed a float

+++ VDTerminal/shared/src/commonMain/kotlin/net/vertexdezign/vdt/Model.kt
@@ -262,7 +262,7 @@ data class Ai(
 
 @Serializable
 data class CruiseControl(
-    val targetSpeed: Int? = null,
+    val targetSpeed: Float? = null,
     val active: Boolean? = null,
 )
